refactored Timeline component to be more flexible
